Added RubyCritic
Provides additional metrics and code quality analysis on top of what is
produced by SimpleCov. This is useful when needing a high level visual
analysis of a project.

It would be nice if this could run on Circle CI but experiencing issues
with the SexpProcessor so opting to ignore RubyCritic on CI for the
time being.

[Details](https://github.com/whitesmith/RubyCritic).