Fixed calculation of empty commits when creating a tag
Commits were being calculated from the beginning of time on large
repositories due to not accounting for the latest Git tag due to the
Git+ upgrade/refactor. We need to use the computed commits and not
whatever is answered back from the repository since this, tagger,
object has all of that knowledge and not Git+ in this situation.

The order operations was changed because the specs would break
otherwise:

- Step 1: Check local version existence.
- Step 2: Check if any commits exist to be tagged.

The private `#exists?` method was renamed to `#local?` to better
reflect that we are checking for existence of a local tag.