Added commit part
Necesssary to encapsulate the presentation of a commit as a line item within a milestone. This includes support for multiple output formats that will eventually be controlled via the Command Line Interface.

Reek was updated to ignore multiple methods in this class since, for now, it's best to keep all of this logic together than split up the functionality further.

Milestone: minor