Added universal renderer
Necessary to provide a primary renderer which can render multiple formats as HTML. This will soon be used within the views for rendering HTML milestones.

Reek has been updated to ignore the unused parameter because it is unable to detect the parameter _is_ being used via the `binding` in order to avoid conflicts with the native keyword.

Milestone: minor