Added container input, spec loader, and memoization
Necessary for CLI input which is a copy of the configuration but allows for mutation based on overrides from the CLI during runtime. This will soon be used by future CLI build actions and commands.

A loader for gem specifications was added for use later throughout this project which means the registration of this gem's specification was updated accordingly.

Additional memoization was added to improve performance but does make the diff more noisy than intended.

Milestone: minor