Added CLI build root action
Necessary to specify or dynamically compute the path when building a milestone. This is the path where milestone documentation will be built for viewing.

The Reek configuration was updated to ignore the control parameter since use of a default is currently a bug with the Option Parser gem.

Milestone: minor