Updated container configuration to use transformers
Necessary to ensure the configuration is populated with safe defaults when applicable. Otherwise, whatever is in the default configuration will be used. Order matters in terms of gem or citation information where gem information is preferred while citations are used as a backup (especially useful in generic projects).

Milestone: major