Added commits versioner
Necessary to analyze unversioned commits, last tag, and determine what the next version should be. This will be used to handle the automatic updating of a version.

Reek was updated to ignore the nested iterator for trailer milestones since breaking this down further wouldn't be of much help so keeping this together instead.

Milestone: minor