Added commits collector
Necessary to collect all untagged commits (i.e. a repository with no tags) or unreleased commits (i.e. a repository with tags and commits which haven't been tagged yet). This is an extraction from the categorizer which will soon be updated to use this object this object instead.

Milestone: minor