Updated publish action to use default version
Necessary to leverage the newly added `Commits::Versioner` instance which can dynamically compute the next version as a default value. This will not only be more informative from the CLI but also allow anyone to override the default if not desired.

The Reek configuration was updated to ignore the control parameter since use of a default is currently a bug with the Option Parser gem.

Milestone: minor