Added build template paths configuration transformer
Necessary to assemble and assign the template paths used by this gem for the Hanami views based on an XDG configuration. If no XDG configuration is found, then the templates provided by this gem will be used instead. This information will soon be used for build purposes and by all views.

Milestone: minor