Added container cache
Necessary to register the cache and corresponding XDG path for use throughout the entire library.

Due to the introduction of this client along with XDG support, the container was updated to support an XDG namespace for managing the cache _and_ configuration. This also required effected objects to be updated accordingly to account for the new namespace.

Milestone: major