Added project label configuration transformer
Necessary to provide a fallback label when the gem specification or citation infromation is not provided. The label produced is a best guess based on current dirctory name which, in some cases, the missing ability to use inflections might not produce the desired result. In those cases, a custom configuration can be used instead.