Updated tag objects to inject input and be called with optional version
Necessary to make these objects more flexible in that they can be initialized with the injected CLI `input` as a version fallback while allowing an optional version _override_ if desired. All versions are type checked for safety.

Reek was updated to ignore the control paramter since the fallback is useful in this situation.

Milestone: major