Updated configuration transformers to simplify default value
Necessary, in most cases, to not check for an empty value and, instead, add the key with a `nil` value so the `nil` value can bubble up to the contract which will validate the type before building the record. Due to this change, some of the specs were updated to reflect the fact you can have a key with a `nil` value. Again, the contract will validate and catch if a `nil` manages to make it through all transformations.

Milestone: patch