Updated commits categorizer to pass range to collector
Necessary to pass this information down to the collector if specified. Otherwise, the fallbacks are used.

Additionally, the configuration (input) is now injected via the constructor which is consistent with most objects in this architecture. Due to this, the processing of commit labels and associated pattern was optimized so this doesn't happen each time commits are categorized.

Milestone: minor