Added build max configuration
Necessary to limit the number of milestones included in a build. By default, this is the most recent changes since the last tag which has been standard behavior since the inception of this project. With the introduction of this new build maximum, you can now specify how far back you want build instead of using only the most recent milestone. The value of the build maximum, in normal situations, is meant to be an integer but a special string (`"all"`) can be provided if all milestones are desired.

Initially, this will be used for syndication purposes with potential for future use with additional build formats.

Milestone: minor