Added build max configuration
Necessary to limit the number of milestones included in a build. By default, this is the most recent changes since the last tag which has been the standard for a while. With the introduction of this new build maximum, you can now specify how far back you want to build instead of using only the most recent tag. The default root folder is now pluralized to accommodate this new functionality as well.

Milestone: major