Added build max CLI action
Necessary to limit how many milestones can be built at once based on the new build max configuration. Initially, this will only be used by the feed format but has potential for other build formats in the future.

Reek has been updated to ignore the control parameter issue because we do want the default as a fallback.

Milestone: minor