Added project URI configuration
Necessary to provide URI defaults which can be calculated from the newly added organization URI and exiting project name. This allows the templates use as needed.

The original project URI has been removed from the context since it's not used.

This will initially be used for syndicated feed purposes but will soon be used in the generation of existing documentation and other metadata too.

Milestone: minor