Added build tail configuration and action
Necessary to define the tail end of the build. Historically, this has been `HEAD` (Git reference) so you can build for all untagged commits since the last known tag. This remains the default so there is no new change in behavior.

What _is_ new is you can now, optionally, cap the build at the last _tag_ created. This is useful for syndicated feed (and potentially web) purposes since you don't want Milestoner to build for the next tag and only build for what is currently known.

Reek was updated to ignore the control parameter since this is behavior we want in this case.

Milestone: minor