Added project logo to web, Markdown, and ASCII doc formats
Necessary to be consistent with the feed format by including the project logo in the generated output. Due to each format having slightly different capabilities, the logo is either centered at the top of the page or included in the article header as best befits the format.

A few additional changes are included in this work:

* For the web format, author profile is now linked so you can click on the author to learn more if desired.
* For ASCII Doc and Markdown, the project label now links to the project home page.
* The stylesheet was updated, now that project logo is used, to ensure all items remain centered and wrap appropriately when on smaller screens.
* Format specific specs were moved out of the show spec and into the respective builder specs to improve organization.

Milestone: minor