Added syndication refinements
Necessary to make dealing with the RSS gem easier since the object API is so poorly designed. In this case, I'm using the `Shared` module to define common _merge_ functionality once and then import that same functionality in all objects that can benefit from the behavior. This will soon be used by the builder (not implemented yet).

Milestone: minor