Updated tag creator, pusher, and publisher to use monads
Necessary to improve fault tolerance of the entire create, push, and publish pipleine. This completes the work started several versions back. 🎉

This also simplifies the creator, pusher, and publisher implementations by not having to cast the version each time. These objects are low-level enough to not have to care about the format of the version which can be moved up to a higher level (i.e. `CLI::Actions::Publish`) so casting to a version only happens once.

Milestone: major