Updated builders to answer monads
Originally, there was enough commonality that `Milestoner::CLI::Commands::Build` could handle logging information about the build process but there is enough subtlety to let `Milestoner::CLI::Commands::Build` be the orchestrator of who builds a particular format instead. This has the added benefit of removing a Reek violation while letting each builder specialize as desired.

There is additional elegance, with this design change, in the fact that the result (monad) bubbles all the way up to each builder which then knows how to deal with either the success or the failure accordingly. There is some duplication in terms of processing the success and/or failure but, at the moment, this is better to keep with the builders rather thant the build command itself.

Milestone: major