Added build tail configuration
Necessary to define the tail end of the build. Historically, this has been `HEAD` (Git reference) so you can build for all untagged commits since the last known tag. This remains the default so there is no new change in behavior. What _is_ new is that you can, optionally, cap the build at the last _tag_ created. This is useful, initially, for syndicated feed purposes -- but will come in handy for the other build formats -- since you don't want Milestoner to build for the next tag and only build for what is currently known.

Milestone: minor