Fixed stream builder to answer content instead of I/O instance
Originally, I wanted to answer an I/O object back which works fine in a test environment but, when used in production, this will usually be `STDOUT` which means you can't rewind to obtain string content. So, to ensure the body is always answered, the builder answers a string of content instead of an I/O object. This allows the `Creator` to build a tag with the content without any additional work.

Milestone: patch