Fixed project icon and logo URLs to be fully optional
Necessary to ensure both the feed and web formats respect custom settings by _not_ including these images if not set. The web format was also updated to fully respect the URI settings in order to be consistent with feed format behavior. Both the web and feed formats are now fully customizable in this regard.

Milestone: patch