Fixed transformers to not compress attributes
Necessary to ensure any values that might be empty to remain empty as this'll be caught during validation. This ensures empty paths, for example, which don't exist on the file system are not accidentally compressed which can happen when you have an empty `tmp/milestones` path which can cause unnecessary Etcher `build_root` validation errors.

This also improves performance since we're not constantly recompressing per each transformer.

Milestone: patch