Updated sanitizer to include common elements and attributes
Necessary to allow more elements in commit messages and notes. This includes providing support for common attributes across all elements.

Due to these changes the implementation needed to be refactored to build the elements and attributes separately. This also resulted in refactoring the spec to be more consistent.

Milestone: minor