Fixed tagger to account for repositories with no tags
Necessary to ensure you can generate release notes for a repository with no existing tags which is a common situation for newly created repositories. This cropped up after this commit was added: 4ebcd578e1e3 (Added commit tagger).

This class is starting to get bottom heavy (i.e. a fair amount of private methods). Namely, the following objects could prove useful:

* A placeholder tag.
* An existing tag.
* A combined tag (i.e. factory) which knows how to build either.

For now, I'm holding off on further extraction but is worth consideration for the next major release.

This also caused the syndication and feed builders to be updated since there will be an array of tags provided now.

Milestone: patch