Added custom view scope builder
Necessary because the default Hanami scope builder is completely broken because you can't use multiple scopes within the same partial which defaults the purpose and entire value of having custom scopes to begin with.

This is also why this scope build is configured within the view to replace the default, broken behavior.

Milestone: minor