Fixed tagger to account for repositories with only a single tag
Now that the `Collector` knows how to deal with references where the first reference is `nil` (i.e. only a single tag exists), we can use this functionality to answers tags with or without commits even if the repository has no commits and tags or only has a single tag.

Milestone: patch