Fixed syndication builder to account for missing commit or tag authors
Previously, when building a feed, you'd get an error when no authors could be found for commits associated with tags. This proved to be too strict so now the builder will check for commit authors and fallback to tag authors. Only when tag and commit authors can't be found will you get an error.

This has the added benefit of bringing this object in line with behavior found with the other build formats when there are no commits (in other words: no activity).

Milestone: patch