Added HTML template past and future versions
Necessary to enhance the viewing experience by providing a link back to the main versions index (root) as well as being able to navigate between previous and next versions. A popover was added to provide keyboard shortcut documentation/help.

The `tag` remains exposed as `tag` instead of `present` in order to remain consistent with terminology used in the other build formats.

Milestone: minor