Added commit created and updated enrichment
Necessary to convert the original `authored_at` and `committed_at` attributes into equivalent `created_at` and `updated_at` instances of `Time` so this can be exposed via the commit part.

All attributes have been alphabetically sorted and RuboCop is updated to ignore collection literal length in the shared context because, for the moment, splitting this up further wouldn't be helpful.

Milestone: minor