Added spinner indicators
Necessary to show a loading indicator when making htmx GET requests which may or may not cause a slight delay in loading new content. This introduces and makes use of the new `spinner` class (and associated keyframe animation) along with a `fragment` class which is only meant to wrap an htmx element and associated spinner.

Relative URLs (i.e `+../+`) are now used for _Previous_ and _Next_ links in order to improve make what is generated via this template easier to embed in other web sites since all versioned directories are siblings to each other. This also makes these links consistent with how the index link works.

Milestone: minor