Fixed horizontal alignment of users within XML templates
Necessary to ensure all user information (avatar + profile) is horizontally centered. The indentation for collaborator and signer lists has been removed so they are consistent with author section information.

Definitely not a fan of inline styles for this but suppling an external style sheet for syndicated feeds isn't compatibility across different feed readers so this is a compromise.

Milestone: minor